refactor(elements): Update ExampleProject
 - Create folders and move files to folders for file hierarchy

 - Fix ViewController indentations and spaces

 - Delete unused functions in TutorialView
